# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for EDF Inscrições
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Registration {
  id: number;
  name: string;
  birthYear: number;
  schoolYear: string;
  className: string;
  activity1?: number | null;
  activity2?: number | null;
  activity3?: number | null;
  activity4?: number | null;
  activity5?: number | null;
  average?: number | null;
  createdAt: string;
}

export interface CreateRegistrationInput {
  name: string;
  birthYear: number;
  schoolYear: string;
  className: string;
}

export interface ScoresInput {
  activity1?: number | null;
  activity2?: number | null;
  activity3?: number | null;
  activity4?: number | null;
  activity5?: number | null;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface LoginResponse {
  success: boolean;
  message: string;
}

export interface MessageResponse {
  message: string;
}

export interface ErrorResponse {
  error: string;
}