# lib/api-zod/src/generated/types/registration.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for EDF Inscrições
 * OpenAPI spec version: 0.1.0
 */

export interface Registration {
  id: number;
  name: string;
  birthYear: number;
  schoolYear: string;
  className: string;
  activity1?: number | null;
  activity2?: number | null;
  activity3?: number | null;
  activity4?: number | null;
  activity5?: number | null;
  average?: number | null;
  createdAt: string;
}